+++ tests/integration_test.rs
@@ -31,6 +31,20 @@ mod tests {
     }
 
     #[test]
+    fn summary_none_with_id() {
+        let test_dir = tempdir().expect("could not create temp directory");
+        let test_dir_path = test_dir.path().to_str().unwrap();
+        let mut cmd = Command::cargo_bin("rtw").unwrap();
+        cmd.arg("-d")
+            .arg(test_dir_path)
+            .arg("summary")
+            .arg("--id")
+            .assert()
+            .success()
+            .stdout("No filtered data found.\n");
+    }
+
+    #[test]
     fn continue_none() {
         let test_dir = tempdir().expect("could not create temp directory");
         let test_dir_path = test_dir.path().to_str().unwrap();
@@ -116,7 +130,7 @@ mod tests {
     }
 
     #[test]
-    fn track_date() {
+    fn track_date_missing_seperator() {
         let test_dir = tempdir().expect("could not create temp directory");
         let test_dir_path = test_dir.path().to_str().unwrap();
         let mut cmd = Command::cargo_bin("rtw").unwrap();
@@ -126,6 +140,173 @@ mod tests {
             .arg("2019-12-25T19:43:00")
             .arg("2019-12-25T19:45:00")
             .arg("foo")
+            .assert()
+            .failure();
+    }
+
+    #[test]
+    fn track_date() {
+        let test_dir = tempdir().expect("could not create temp directory");
+        let test_dir_path = test_dir.path().to_str().unwrap();
+        let mut cmd = Command::cargo_bin("rtw").unwrap();
+        cmd.arg("-d")
+            .arg(test_dir_path)
+            .arg("track")
+            .arg("2019-12-25T19:43:00")
+            .arg("-")
+            .arg("2019-12-25T19:45:00")
+            .arg("foo")
+            .assert()
+            .success();
+    }
+
+    #[test]
+    fn track_relative() {
+        let test_dir = tempdir().expect("could not create temp directory");
+        let test_dir_path = test_dir.path().to_str().unwrap();
+        let mut cmd = Command::cargo_bin("rtw").unwrap();
+        cmd.arg("-d")
+            .arg(test_dir_path)
+            .arg("track")
+            .arg("10 min ago")
+            .arg("-")
+            .arg("5 min ago")
+            .arg("foo")
+            .assert()
+            .success();
+    }
+
+    #[test]
+    fn track_relative_missing_end() {
+        let test_dir = tempdir().expect("could not create temp directory");
+        let test_dir_path = test_dir.path().to_str().unwrap();
+        let mut cmd = Command::cargo_bin("rtw").unwrap();
+        cmd.arg("-d")
+            .arg(test_dir_path)
+            .arg("track")
+            .arg("10 min ago")
+            .arg("-")
+            .arg("foo")
+            .assert()
+            .success();
+    }
+
+    #[test]
+    fn track_relative_missing_start_and_end() {
+        let test_dir = tempdir().expect("could not create temp directory");
+        let test_dir_path = test_dir.path().to_str().unwrap();
+        let mut cmd = Command::cargo_bin("rtw").unwrap();
+        cmd.arg("-d")
+            .arg(test_dir_path)
+            .arg("track")
+            .arg("-")
+            .arg("foo")
+            .assert()
+            .failure();
+    }
+
+    #[test]
+    fn start_nothing_now() {
+        let test_dir = tempdir().expect("could not create temp directory");
+        let test_dir_path = test_dir.path().to_str().unwrap();
+        let mut cmd = Command::cargo_bin("rtw").unwrap();
+        cmd.arg("-d")
+            .arg(test_dir_path)
+            .arg("start")
+            .assert()
+            .failure();
+    }
+
+    #[test]
+    fn start_nothing_15min_ago() {
+        let test_dir = tempdir().expect("could not create temp directory");
+        let test_dir_path = test_dir.path().to_str().unwrap();
+        let mut cmd = Command::cargo_bin("rtw").unwrap();
+        cmd.arg("-d")
+            .arg(test_dir_path)
+            .arg("start")
+            .arg("15min")
+            .arg("ago")
+            .assert()
+            .failure();
+    }
+
+    #[test]
+    fn start_foo_15min_ago() {
+        let test_dir = tempdir().expect("could not create temp directory");
+        let test_dir_path = test_dir.path().to_str().unwrap();
+        let mut cmd = Command::cargo_bin("rtw").unwrap();
+        cmd.arg("-d")
+            .arg(test_dir_path)
+            .arg("start")
+            .arg("15min")
+            .arg("ago")
+            .arg("foo")
+            .assert()
+            .success();
+    }
+
+    #[test]
+    fn start_foo_today_at_9() {
+        let test_dir = tempdir().expect("could not create temp directory");
+        let test_dir_path = test_dir.path().to_str().unwrap();
+        let mut cmd = Command::cargo_bin("rtw").unwrap();
+        cmd.arg("-d")
+            .arg(test_dir_path)
+            .arg("start")
+            .arg("09:00")
+            .arg("foo")
+            .assert()
+            .success();
+    }
+
+    #[test]
+    fn start_foo_on_datetime() {
+        let test_dir = tempdir().expect("could not create temp directory");
+        let test_dir_path = test_dir.path().to_str().unwrap();
+        let mut cmd = Command::cargo_bin("rtw").unwrap();
+        cmd.arg("-d")
+            .arg(test_dir_path)
+            .arg("start")
+            .arg("2019-12-24T19:43:00")
+            .arg("foo")
+            .assert()
+            .success();
+    }
+
+    #[test]
+    fn stop_nothing_now() {
+        let test_dir = tempdir().expect("could not create temp directory");
+        let test_dir_path = test_dir.path().to_str().unwrap();
+        let mut cmd = Command::cargo_bin("rtw").unwrap();
+        cmd.arg("-d")
+            .arg(test_dir_path)
+            .arg("stop")
+            .assert()
+            .success();
+    }
+
+    #[test]
+    fn stop_foo_5min_ago() {
+        let test_dir = tempdir().expect("could not create temp directory");
+        let test_dir_path = test_dir.path().to_str().unwrap();
+        let mut cmd = Command::cargo_bin("rtw").unwrap();
+        cmd.arg("-d")
+            .arg(test_dir_path)
+            .arg("start")
+            .arg("10")
+            .arg("min")
+            .arg("ago")
+            .arg("foo")
+            .assert()
+            .success();
+        let mut cmd = Command::cargo_bin("rtw").unwrap();
+        cmd.arg("-d")
+            .arg(test_dir_path)
+            .arg("stop")
+            .arg("5")
+            .arg("min")
+            .arg("ago")
             .assert()
             .success();
     }